PlayerContextProvider separated from AudioPlayer.
